api/requirements/base.txt
psycopg2-binary==2.9.1
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cfgv==3.3.1
channels==4.0.0
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
django-mailersend==0.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.12.2
frozenlist==1.3.3
identify==2.5.24
idna==3.4
mailersend==0.5.1
multidict==6.0.4
nodeenv==1.8.0
Pillow==9.5.0
platformdirs==3.8.0
pre-commit==2.21.0
promise==2.3
PyJWT==2.7.0
pyotp==2.8.0
pyserial==3.5
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
twilio==8.3.0
typing_extensions==4.6.3
urllib3==2.0.3
virtualenv==20.23.1
yarl==1.9.2